touristes/requirements.txt
blinker
click
contourpy
cycler
Flask
Flask-Login
Flask-SQLAlchemy
fonttools
greenlet
itsdangerous
Jinja2
joblib
kiwisolver
MarkupSafe
matplotlib
numpy
packaging
pandas
pillow
playwright
psycopg2-binary
pyee
pyparsing
python-dateutil
python-dotenv
pytz
scikit-learn
scipy
seaborn
six
SQLAlchemy
threadpoolctl
typing_extensions
tzdata
Werkzeug
xgboost
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
